Fix all high/medium/low security vulnerabilities; set SSH timeout to 4s
C-1  crypto: embed per-blob random 16-byte PBKDF2 salt in encrypted blobs
H-1  window: add CRITICAL_SECTION to guard conn_result/conn_error between threads
H-2  ssh_session: replace gethostbyname with thread-safe getaddrinfo (IPv4+IPv6)
H-3  window: SecureZeroMemory(conn_profile.password) after auth completes
H-4  knownhosts: pass and map actual key type (RSA/ECDSA/Ed25519) in add()
M-1  crypto: reject plaintext > 64 KB to prevent size_t overflow in b64
M-2  crypto: check RegQueryValueExA return value; null-terminate on failure only
M-3  loader: reject config files > 1 MB in read_file()
M-4  loader: write to .tmp then rename for atomic config save
M-5  window: SecureZeroMemory(ctx.buf) on passphrase dialog cancellation too
M-6  window: validate log_format against whitelist before strftime
M-7  ssh_session: validate port range 1-65535 before connect
M-8  window: resolve config.json to absolute exe-relative path at startup
I-1  ssh_session: 4-second TCP connect timeout via non-blocking socket + select
I-3  window: show error dialog and clean up when tab limit is reached
L-1  crypto/loader/ssh_session: use secure_zero (volatile) instead of memset
L-2  crypto: guard against pt_len > 64 KB before int cast in EVP_EncryptUpdate
L-3  loader: escape control characters < 0x20 as \uXXXX in JSON output
L-4  parser: add explicit bounds check before CSI param array access
L-5  ssh_io: change total_read accumulator from int to size_t

+++ src/config/ssh_io.c
@@ -25,7 +25,7 @@ int ssh_io_poll(SSHChannel *channel, Terminal *term, FILE *log_file) {
     if (!channel || !term) return -1;
 
     char buf[4096];
-    int total_read = 0;
+    size_t total_read = 0;  /* L-5: use size_t to avoid int overflow */
     int loops = 0;
     const int MAX_LOOPS = 10;
 
@@ -37,7 +37,7 @@ int ssh_io_poll(SSHChannel *channel, Terminal *term, FILE *log_file) {
         if (rc > 0) {
             term_process(term, buf, (size_t)rc);
             log_chunk(log_file, buf, (size_t)rc);
-            total_read += (int)rc;
+            total_read += (size_t)rc;
             work_done = 1;
         } else if (rc == LIBSSH2_ERROR_EAGAIN || rc == 0) {
             if (libssh2_channel_eof(channel->channel)) return -2; /* EOF */
@@ -50,7 +50,7 @@ int ssh_io_poll(SSHChannel *channel, Terminal *term, FILE *log_file) {
         if (rc > 0) {
             term_process(term, buf, (size_t)rc);
             log_chunk(log_file, buf, (size_t)rc);
-            total_read += (int)rc;
+            total_read += (size_t)rc;
             work_done = 1;
         }
